Rename widget folders and IDs to match current URLs
