+++ src/applications/static-pages/health-care-manage-benefits/secure-messaging-page/components/App/index.unit.spec.js
