+++ src/applications/static-pages/health-care-manage-benefits/secure-messaging-page/components/AuthContent/index.js
