+++ src/applications/static-pages/health-care-manage-benefits/view-test-and-lab-results-page/components/App/index.unit.spec.js
