+++ src/applications/static-pages/health-care-manage-benefits/view-test-and-lab-results-page/components/AuthContent/index.unit.spec.js
