+++ src/applications/static-pages/health-care-manage-benefits/view-test-and-lab-results-page/index.js
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 export default (store, widgetType) => {
   const root = document.querySelector(`[data-widget-type="${widgetType}"]`);
   if (root) {
-    import(/* webpackChunkName: "view-test-lab-results-page" */
+    import(/* webpackChunkName: "view-test-and-lab-results-page" */
     './components/App').then(module => {
       const App = module.default;
       ReactDOM.render(
